Completed Week 2 Task 1 - Acyclicity
Completed this with keeping track of current elements in the call stack
by passing a copy of the elements visited to further recursive call in
this call frame.

I am not confident if the solution that I did was the best possible solution,
I will look for better implementation for the same problem and update
if found something interesting.